Moved obsolete code and updated branch
