+++ resources/scrap.py
@@ -25,261 +25,6 @@ import time
 from utils import *
 from objects import *
 
-# -------------------------------------------------------------------------------------------------
-# We support online an offline scrapers.
-# Note that this module does not depend on Kodi stuff at all, and can be
-# called externally from console Python scripts for testing of the scrapers.
-# -------------------------------------------------------------------------------------------------
-
-# --- "GLOBALS" -----------------------------------------------------------------------------------
-DEBUG_SCRAPERS = 1
-
-#--------------------------------------------------------------------------------------------------
-# Base class for all scrapers
-#--------------------------------------------------------------------------------------------------
-class LegacyScraper:
-    # Short name to refer to object in code
-    name = ''
-
-    # This function is called when the user wants to search a whole list of games.
-    #   search_string   Online scrapers use this
-    #   rom_base_noext  Offline scrapers and online MAME require the unmodified ROM name
-    #   platform        AEL platform name. Will be translated to scraper internal name
-    #
-    # Returns:
-    #   results = [game, game, ... ]
-    #   game = {
-    #       'id' : str,             String that allows to obtain the game URL (ID number) or the URL itself.
-    #       'display_name' : str,   Scraped name of the ROM/Launcher/Game.
-    #       ...
-    #   }
-    #   game dictionary may have more fields depending on the scraper (which are not used outside that scraper)
-    def get_search(self, search_string, rom_base_noext, platform):
-        raise NotImplementedError('Subclass must implement get_search() abstract method')
-
-# -------------------------------------------------------------------------------------------------
-# Metadata scrapers base class
-# All scrapers (offline or online) must implement the abstract methods.
-# -------------------------------------------------------------------------------------------------
-class Scraper_Metadata(LegacyScraper):
-    def new_gamedata_dic(self):
-        gamedata = {
-            'title'     : '',
-            'year'      : '',
-            'genre'     : '',
-            'developer' : '',
-            'nplayers'  : '',
-            'esrb'      : '',
-            'plot'      : ''
-        }
-
-        return gamedata
-
-    # Offline scrapers need to know the plugin installation directory.
-    # For offline scrapers just pass.
-    def set_addon_dir(self, plugin_dir):
-        raise NotImplementedError('Subclass must implement set_addon_dir() abstract method')
-
-    # This is called after get_games_search() to get metadata of a particular ROM.
-    # game is a dictionary from the dictionary list returned by get_game_search()
-    # get_game_search() is usually common code for the online scrapers.
-    #
-    # Mandatory fields returned:
-    #   gamedata dictionary created by new_gamedata_dic()
-    def get_metadata(self, game):
-        raise NotImplementedError('Subclass must implement get_metadata() abstract method')
-
-# --- Asset scrapers ------------------------------------------------------------------------------
-# All asset scrapers are online scrapers.
-# -------------------------------------------------------------------------------------------------
-class Scraper_Asset(LegacyScraper):
-    # If scraper needs additional configuration then call this function.
-    def set_options(self, region, imgsize):
-        raise NotImplementedError('Subclass must implement set_options() abstract method')
-
-    # Returns True if scraper supports the asset, False otherwise.
-    def supports_asset(self, asset_kind):
-        raise NotImplementedError('Subclass must implement supports_asset() abstract method')
-        
-    # Obtain a set of images of the given kind, based on a previous search with get_search()
-    #
-    # Returns:
-    #   images = [image_dic, image_dic, ... ]
-    #   image_dic = {
-    #       'name' : str,   Name of the image (e.g., 'Boxfront 1')
-    #       'id'   : str,   String that allows to obtain the game URL (ID number) or the URL itself.
-    #       'URL'  : str    URL of a thumb to display the image. Some websites have small size
-    #                           images for preview. It could be the URL of the full size image itself.
-    #       ...    : ...    Scrapers may add additional fields for image_dic dictionary if needed.
-    #   }
-    def get_images(self, game, asset_kind):
-        raise NotImplementedError('Subclass must implement get_images() abstract method')
-
-    # Resolves the full size image URL. id is the string returned by get_images()
-    # Returns:
-    #   (URL,            Full URL of the full size image.
-    #    img_extension)  Extension of the image.
-    def resolve_image_URL(self, image_dic):
-        raise NotImplementedError('Subclass must implement resolve_image_URL() abstract method')
-
-#--------------------------------------------------------------------------------------------------
-# Instantiate scraper objects
-#--------------------------------------------------------------------------------------------------
-from scrap_metadata import *
-from scrap_asset import *
-
-# --- Metadata scraper objects ---
-Offline_meta_obj    = metadata_Offline()
-TheGamesDB_meta_obj = metadata_TheGamesDB()
-GameFAQs_meta_obj   = metadata_GameFAQs()
-MobyGames_meta_obj  = metadata_MobyGames()
-ArcadeDB_meta_obj   = metadata_ArcadeDB()
-
-# --- Metadata scrapers ---
-# >> This list MUST match the settings configuration in settings.xml or bad things will happen.
-scrapers_metadata = [
-    Offline_meta_obj, TheGamesDB_meta_obj, GameFAQs_meta_obj, MobyGames_meta_obj
-]
-
-scrapers_metadata_MAME = [
-    Offline_meta_obj, TheGamesDB_meta_obj, GameFAQs_meta_obj, MobyGames_meta_obj, ArcadeDB_meta_obj
-]
-
-# --- Asset scraper objects ---
-# >> There is only one instantiated object for every scraper. Scraper objects cache all possible
-# >> requests made to the scraper to save bandwidth and increase speed.
-NULL_obj       = asset_NULL()
-TheGamesDB_obj = asset_TheGamesDB()
-GameFAQs_obj   = asset_GameFAQs()
-MobyGames_obj  = asset_MobyGames()
-ArcadeDB_obj   = asset_ArcadeDB()
-
-# --- Asset scrapers ---
-# >> This list is used in _gui_edit_asset()
-scrapers_asset = [TheGamesDB_obj, GameFAQs_obj, MobyGames_obj, ArcadeDB_obj]
-
-# >> These lists MUST match the settings configuration in settings.xml or bad things will happen.
-# >> These lists are used in the ROM Scanner.
-# >> Boxfront  -> Cabinet
-# >> Boxback   -> CPanel
-# >> Cartridge -> PCB
-# >> Banner    -> Marquee
-scrapers_title          = [ NULL_obj, TheGamesDB_obj, GameFAQs_obj, MobyGames_obj               ]
-scrapers_snap           = [ NULL_obj, TheGamesDB_obj, GameFAQs_obj, MobyGames_obj               ]
-scrapers_boxfront       = [ NULL_obj, TheGamesDB_obj, GameFAQs_obj, MobyGames_obj               ]
-scrapers_boxback        = [ NULL_obj, TheGamesDB_obj, GameFAQs_obj, MobyGames_obj               ]
-scrapers_cartridge      = [ NULL_obj,                               MobyGames_obj               ]
-scrapers_fanart         = [ NULL_obj, TheGamesDB_obj                                            ]
-scrapers_banner         = [ NULL_obj, TheGamesDB_obj                                            ]
-scrapers_clearlogo      = [ NULL_obj, TheGamesDB_obj                                            ]
-
-# >> MAME scrapers
-scrapers_title_MAME     = [ NULL_obj, TheGamesDB_obj, GameFAQs_obj, MobyGames_obj, ArcadeDB_obj ]
-scrapers_snap_MAME      = [ NULL_obj, TheGamesDB_obj, GameFAQs_obj, MobyGames_obj, ArcadeDB_obj ]
-scrapers_cabinet_MAME   = [ NULL_obj, TheGamesDB_obj, GameFAQs_obj, MobyGames_obj, ArcadeDB_obj ]
-scrapers_cpanel_MAME    = [ NULL_obj, TheGamesDB_obj, GameFAQs_obj, MobyGames_obj, ArcadeDB_obj ]
-scrapers_pcb_MAME       = [ NULL_obj,                               MobyGames_obj, ArcadeDB_obj ]
-scrapers_fanart_MAME    = [ NULL_obj, TheGamesDB_obj                                            ]
-scrapers_marquee_MAME   = [ NULL_obj, TheGamesDB_obj,                              ArcadeDB_obj ]
-scrapers_clearlogo_MAME = [ NULL_obj, TheGamesDB_obj,                              ArcadeDB_obj ]
-scrapers_flyer_MAME     = [ NULL_obj,                                              ArcadeDB_obj ]
-
-# #################################################################################################
-# #################################################################################################
-# Scraper factories.
-# #################################################################################################
-# #################################################################################################
-
-def getScraper(asset_kind, settings, useMame = False):
-    if useMame:
-        return getMameScraper(asset_kind, settings)
-
-    key = ASSET_SETTING_KEYS[asset_kind]
-
-    if key == '':
-        return None
-
-    if key not in settings:
-        log_warning("Scraper with key {} not set in settings".format(key))
-        return None
-
-    scraper_index = settings[key]
-
-    if asset_kind == ASSET_TITLE:
-        return scrapers_title[scraper_index]
-
-    if asset_kind == ASSET_SNAP:
-        return scrapers_snap[scraper_index]
-
-    if asset_kind == ASSET_BOXFRONT:
-        return scrapers_boxfront[scraper_index]
-
-    if asset_kind == ASSET_BOXBACK:
-        return scrapers_boxback[scraper_index]
-
-    if asset_kind == ASSET_CARTRIDGE:
-        return scrapers_cartridge[scraper_index]
-
-    if asset_kind == ASSET_FANART:
-        return scrapers_fanart[scraper_index]
-
-    if asset_kind == ASSET_BANNER:
-        return scrapers_banner[scraper_index]
-
-    if asset_kind == ASSET_CLEARLOGO:
-        return scrapers_clearlogo[scraper_index]
-
-    # >> Flyers only supported by ArcadeDB (for MAME). If platform is not MAME then deactivate
-    # >> this scraper.
-    # >> Map (not supported yet, use a null scraper)
-    # >> Manual (not supported yet, use a null scraper)
-    # >> Trailer (not supported yet, use a null scraper)
-    return NULL_obj
-
-def getMameScraper(asset_kind, settings):
-    key = MAME_ASSET_SETTING_KEYS[asset_kind]
-
-    if key == '':
-        return None
-
-    if key not in settings:
-        log_warning("Scraper with key {} not set in settings".format(key))
-        return None
-
-    scraper_index = settings[key]
-
-    if asset_kind == ASSET_TITLE:
-        return scrapers_title_MAME[scraper_index]
-
-    if asset_kind == ASSET_SNAP:
-        return scrapers_snap_MAME[scraper_index]
-
-    if asset_kind == ASSET_BOXFRONT:
-        return scrapers_cabinet_MAME[scraper_index]
-
-    if asset_kind == ASSET_BOXBACK:
-        return scrapers_cpanel_MAME[scraper_index]
-
-    if asset_kind == ASSET_CARTRIDGE:
-        return scrapers_pcb_MAME[scraper_index]
-
-    if asset_kind == ASSET_FANART:
-        return scrapers_fanart_MAME[scraper_index]
-
-    if asset_kind == ASSET_BANNER:
-        return scrapers_marquee_MAME[scraper_index]
-
-    if asset_kind == ASSET_CLEARLOGO:
-        return scrapers_clearlogo_MAME[scraper_index]
-
-    if asset_kind == ASSET_FLYER:
-        return scrapers_flyer_MAME[scraper_index]
-
-    # >> Map (not supported yet, use a null scraper)
-    # >> Manual (not supported yet, use a null scraper)
-    # >> Trailer (not supported yet, use a null scraper)
-    return NULL_obj
-
 class ScraperFactory(KodiProgressDialogStrategy):
     def __init__(self, PATHS, settings):
         self.settings = settings
@@ -467,10 +212,6 @@ class ScraperSettings(object):
             settings['scan_ignore_scrap_title'] if 'scan_ignore_scrap_title' in settings else False,
             settings['scan_clean_tags'] if 'scan_clean_tags' in settings else True)
 
-class ScrapeResult(object):
-    metada_scraped  = False
-    assets_scraped  = []
-
 class Scraper(object):
     __metaclass__ = abc.ABCMeta
     
@@ -587,48 +328,7 @@ class Scraper(object):
             return self.fallbackScraper.scrape_asset(search_term, rom_path, rom)
         
         return scraper_applied
-
-    def scrape(self, search_term, romPath, rom):
-                
-        results = self._get_candidates(search_term, romPath, rom)
-        log_debug('Scraper \'{0}\' found {1} result/s'.format(self.getName(), len(results)))
-
-        if not results or len(results) == 0:
-            log_verb('Scraper \'{0}\' found no games after searching.'.format(self.getName()))
-            
-            if self.fallbackScraper:
-                return self.fallbackScraper.scrape(search_term, romPath, rom)
-
-            return False
-
-        selected_candidate = 0
-        if len(results) > 1 and self.scraper_settings.metadata_scraping_mode == 0:
-            log_debug('Metadata semi-automatic scraping')
-            # >> Close progress dialog (and check it was not canceled)
-            # ?? todo again
-
-            # >> Display corresponding game list found so user choses
-            dialog = xbmcgui.Dialog()
-            rom_name_list = []
-            for game in results: 
-                rom_name_list.append(game['display_name'])
-
-            selected_candidate = dialog.select('Select game for ROM {0}'.format(romPath.getBase_noext()), rom_name_list)
-            if selected_candidate < 0: selected_candidate = 0
-
-            # >> Open progress dialog again
-            # ?? todo again
         
-        self._loadCandidate(results[selected_candidate], romPath, rom)
-                
-        scraper_applied = self._apply_candidate(romPath, rom)
-                
-        if not scraper_applied and self.fallbackScraper is not None:
-            log_verb('Scraper \'{0}\' did not get the correct data. Using fallback scraping method: {1}.'.format(self.getName(), self.fallbackScraper.getName()))
-            scraper_applied = self.fallbackScraper.scrape(search_term, romPath, rom)
-        
-        return scraper_applied
-    
     @abc.abstractmethod
     def getName(self):
         return ''
@@ -638,17 +338,6 @@ class Scraper(object):
     @abc.abstractmethod
     def _get_candidates(self, search_term, romPath, rom):
         return []
-    
-    def _loadCandidate(self, candidate, romPath, rom):
-        
-        self.gamedata = self._new_gamedata_dic()
-        
-        if self.scrape_metadata:
-            self._load_metadata(candidate, romPath, rom)
-
-        if self.assets_to_scrape is not None and len(self.assets_to_scrape) > 0:
-            self._load_assets(candidate, romPath, rom)
-    
             
     @abc.abstractmethod
     def _load_metadata(self, candidate, romPath, rom, game_data):
